requirements.txt
transformers
transformers[sentencepiece]
torch
matplotlib
datasets
sacrebleu
rouge_score
pandas
nltk
tqdm
pyyaml
py7zr
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
box
# -e .